changed health and resource from int to double

+++ src/main/java/de/raidcraft/mobs/api/AbstractMob.java
@@ -110,7 +110,7 @@ public abstract class AbstractMob extends AbstractSkilledCharacter<Mob> implemen
     }
 
     @Override
-    public int getDefaultHealth() {
+    public double getDefaultHealth() {
 
         return getMaxHealth();
     }
